+++ web/packages/plugins/ui/sima-task/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@dmt/edit-sima-task",
+  "name": "@dmt/sima-task",
   "license": "MIT",
   "version": "1.0.0",
   "main": "src/index.tsx",
