+++ web/packages/plugins/ui/sima-task/src/index.tsx
@@ -2,6 +2,7 @@ import * as React from 'react'
 
 import { DmtPluginType } from '@dmt/common'
 import { EditSimaTask } from './EditSimaTask'
+import { ViewSimaTask } from './ViewSimaTask'
 
 export const plugins: any = [
   {
@@ -9,6 +10,13 @@ export const plugins: any = [
     pluginType: DmtPluginType.UI,
     content: {
       component: EditSimaTask,
+    },
+  },
+  {
+    pluginName: 'view-sima-task',
+    pluginType: DmtPluginType.UI,
+    content: {
+      component: ViewSimaTask,
     },
   },
 ]
