+++ web/packages/plugins/ui/sima-task/tsconfig.json
